new modifiers for item properties: - APs to reload - APs to reload manually - max distance for messy death - gun handling - best laser range
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@6472 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Interface Items.cpp
@@ -7729,7 +7729,7 @@ void DeleteItemDescriptionBox( )
 	{
 		if (gpAttachSoldier)
 		{
-			UINT8 ubAPCost = 0;
+			INT16 ubAPCost = 0;
 
 			// check for change in attachments
 			//unsigned int originalSize = gOriginalAttachments.size();
